+++ .gitignore
@@ -1,0 +1,2 @@
+.vercel
+node_modules
