--- index.html
@@ -1,1 +1,0 @@
-The embed creation preview page is a work in progress.
