+++ vercel.json
@@ -1,0 +1,6 @@
+{
+    "routes": [
+      { "src": "/create", "dest": "/api/create" },
+      { "src": "/([^/]+)", "dest": "/api" }
+    ]
+  }
